fix(gold): add getOad() to gold.hp; faster coin gravity; Scalar cast
gold.hp was missing the INLINE getOad() accessor, so every build since
GoldValue was added to Collision failed silently (set -e aborted before
linking), leaving the May-21 binary in place with none of the coin fixes.

Changes:
- gold.hp: include gold.ht + INLINE const _Gold* getOad() const
- gold.cc: Scalar((float)GoldValue) to resolve int32 overload ambiguity
- blender_create_smb.py: wf_Falling Acceleration 4.0→12.0 so coin arc
  completes in ~1 s (peak ~1.5 m) instead of 3 s; coin lands near player
  with ~1 s TTL remaining for pickup
- TODO: document Scalar(int32) ambiguity for future fix

+++ wfsource/source/game/gold.cc
@@ -9,7 +9,6 @@
 // See TODO.md (BUILD / TOOLCHAIN) for the broken-.ht-codegen tracking.
 //============================================================================
 
-#include <oas/gold.ht>					// _Gold oad structure
 #include "gold.hp"
 #include "actor.hp"						// object enumeration (Player_KIND)
 #include "level.hp"						// theLevel->SetPendingRemove
@@ -66,7 +65,7 @@ Gold::Collision(PhysicalObject& other, const Vector3& normal)
 		if (player.kind() == Actor::Player_KIND)
 		{
 			Mailboxes& pm = player.GetMailboxes();
-			pm.WriteMailbox(EMAILBOX_GOLD, pm.ReadMailbox(EMAILBOX_GOLD) + Scalar(getOad()->GoldValue));
+			pm.WriteMailbox(EMAILBOX_GOLD, pm.ReadMailbox(EMAILBOX_GOLD) + Scalar((float)getOad()->GoldValue));
 			theLevel->SetPendingRemove(this);
 		}
 	}
